refactor: metric explorer tooltip component
