+++ packages/frontend/src/features/metricsCatalog/components/visualization/MetricsVisualization.tsx
@@ -1,21 +1,17 @@
 import {
     applyCustomFormat,
     capitalize,
-    ComparisonFormatTypes,
     friendlyName,
     getCustomFormat,
     MetricExplorerComparison,
-    TimeFrames,
-    type MetricExploreDataPointWithDateValue,
     type MetricExplorerDateRange,
     type MetricExplorerQuery,
     type MetricsExplorerQueryResults,
     type TimeDimensionConfig,
+    type TimeFrames,
 } from '@lightdash/common';
 import {
-    Badge,
     Box,
-    Divider,
     Flex,
     Group,
     LoadingOverlay,
@@ -23,9 +19,7 @@ import {
     Text,
     Tooltip,
     useMantineTheme,
-    type DefaultMantineColor,
 } from '@mantine/core';
-import { IconLineDashed, IconMinus } from '@tabler/icons-react';
 import { scaleTime } from 'd3-scale';
 import {
     timeDay,
@@ -37,7 +31,6 @@ import {
     timeYear,
 } from 'd3-time';
 import dayjs from 'dayjs';
-import { uniqBy } from 'lodash';
 import { useCallback, useEffect, useMemo, useState, type FC } from 'react';
 import {
     CartesianGrid,
@@ -50,24 +43,14 @@ import {
     Tooltip as RechartsTooltip,
     XAxis,
     YAxis,
-    type TooltipProps as RechartsTooltipProps,
 } from 'recharts';
-import {
-    type NameType,
-    type ValueType,
-} from 'recharts/types/component/DefaultTooltipContent';
-import MantineIcon from '../../../../components/common/MantineIcon';
-import { calculateComparisonValue } from '../../../../hooks/useBigNumberConfig';
 import { useAppSelector } from '../../../sqlRunner/store/hooks';
 import { useDynamicYAxisWidth } from '../../hooks/useDynamicYAxisWidth';
-import {
-    getGranularityLabel,
-    getGranularitySublabel,
-    is5YearDateRange,
-} from '../../utils/metricPeekDate';
+import { is5YearDateRange } from '../../utils/metricPeekDate';
 import { MetricPeekDatePicker } from '../MetricPeekDatePicker';
 import { MetricsVisualizationEmptyState } from '../MetricsVisualizationEmptyState';
 import { MetricExploreLegend } from './MetricExploreLegend';
+import { MetricExploreTooltip } from './MetricExploreTooltip';
 import { TimeDimensionPicker } from './TimeDimensionPicker';
 import { FORMATS } from './types';
 // REMOVE COMMENTS TO ENABLE CHART ZOOM
@@ -91,333 +74,6 @@ const tickFormatter = (date: Date) => {
             ? FORMATS.month
             : FORMATS.year
     )(date);
-};
-
-type RechartsTooltipPropsPayload = NonNullable<
-    RechartsTooltipProps<ValueType, NameType>['payload']
->[number];
-
-interface CustomTooltipPropsPayload extends RechartsTooltipPropsPayload {
-    payload: MetricExploreDataPointWithDateValue;
-}
-
-interface CustomTooltipProps extends RechartsTooltipProps<ValueType, NameType> {
-    comparison: MetricExplorerQuery;
-    granularity: TimeFrames | undefined;
-    is5YearDateRangePreset: boolean;
-    payload?: CustomTooltipPropsPayload[];
-    dateRange?: MetricExplorerDateRange;
-}
-
-const SquareBadge = ({ color }: { color?: DefaultMantineColor }) => (
-    <Box
-        sx={{
-            width: 10,
-            height: 10,
-            borderRadius: 2,
-            backgroundColor: color ?? 'indigo.6',
-        }}
-    />
-);
-
-const CustomTooltipPayloadEntry = ({
-    entry,
-    color,
-    comparison,
-    date,
-    is5YearDateRangePreset,
-    dateRange,
-}: {
-    entry: CustomTooltipPropsPayload;
-    color?: DefaultMantineColor;
-    comparison: MetricExplorerQuery;
-    date: string | undefined;
-    is5YearDateRangePreset: boolean;
-    dateRange?: MetricExplorerDateRange;
-}) => {
-    const entryData = useMemo(() => {
-        if (!entry.name) {
-            return null;
-        }
-
-        const isCompareMetric = entry.name === 'compareMetric';
-        return isCompareMetric
-            ? entry.payload.compareMetric
-            : entry.payload.metric;
-    }, [entry]);
-
-    if (!entryData) {
-        return null;
-    }
-
-    if (
-        comparison.comparison === MetricExplorerComparison.NONE &&
-        comparison.segmentDimension === null
-    ) {
-        return (
-            <Badge
-                variant="light"
-                color="indigo"
-                radius="md"
-                sx={(theme) => ({
-                    border: `1px solid ${theme.colors.indigo[1]}`,
-                })}
-                h={24}
-            >
-                {entryData.formatted}
-            </Badge>
-        );
-    }
-
-    if (comparison.comparison === MetricExplorerComparison.PREVIOUS_PERIOD) {
-        const currentPeriodYear = dateRange ? dayjs(dateRange[1]).year() : null;
-        const startYear = dateRange ? dayjs(dateRange[0]).year() : null;
-
-        let label = getGranularitySublabel(entry.name, date);
-
-        if (is5YearDateRangePreset && startYear && currentPeriodYear) {
-            label =
-                entry.name === 'metric'
-                    ? `${startYear}-${currentPeriodYear}`
-                    : `${startYear - 1}-${currentPeriodYear - 1}`;
-        }
-
-        return (
-            <>
-                <Group position="apart">
-                    <Group spacing={4}>
-                        <MantineIcon
-                            color={color ?? 'indigo.6'}
-                            icon={
-                                entry.name === 'metric'
-                                    ? IconMinus
-                                    : IconLineDashed
-                            }
-                        />
-                        <Text c="gray.8" fz={13} fw={500}>
-                            {label}
-                        </Text>
-                    </Group>
-
-                    <Badge
-                        variant="light"
-                        color="gray.7"
-                        radius="md"
-                        h={24}
-                        sx={(theme) => ({
-                            border: `1px solid ${theme.colors.gray[1]}`,
-                            fontFeatureSettings: '"tnum"',
-                        })}
-                    >
-                        {entryData.formatted}
-                    </Badge>
-                </Group>
-            </>
-        );
-    }
-
-    if (
-        comparison.comparison === MetricExplorerComparison.NONE &&
-        comparison.segmentDimension !== null
-    ) {
-        return (
-            <Group position="apart">
-                <Group spacing={4}>
-                    <SquareBadge color={color} />
-                    <Text c="gray.8" fz={13} fw={500}>
-                        {entryData.label}
-                    </Text>
-                </Group>
-
-                <Badge
-                    variant="light"
-                    color="gray"
-                    radius="md"
-                    h={24}
-                    sx={(theme) => ({
-                        border: `1px solid ${theme.colors.gray[1]}`,
-                        color: theme.colors.gray[7],
-                    })}
-                >
-                    {entryData.formatted}
-                </Badge>
-            </Group>
-        );
-    }
-
-    return (
-        <Group position="apart">
-            <Group spacing={4}>
-                <MantineIcon
-                    color={color ?? 'indigo.6'}
-                    icon={entry.name === 'metric' ? IconMinus : IconLineDashed}
-                />
-                <Text c="gray.8" fz={13} fw={500}>
-                    {entryData.label}
-                </Text>
-            </Group>
-
-            <Badge
-                variant="light"
-                color="gray.7"
-                radius="md"
-                h={24}
-                sx={(theme) => ({
-                    border: `1px solid ${theme.colors.gray[1]}`,
-                })}
-            >
-                {entryData.formatted}
-            </Badge>
-        </Group>
-    );
-};
-
-function getUniqueEntryKey(entry: CustomTooltipPropsPayload) {
-    return `${entry.name}_${entry.payload.segment}_${entry.payload.dateValue}_${entry.payload.metric.value}`;
-}
-
-const PercentageChangeFooter: FC<{
-    uniqueEntries: CustomTooltipPropsPayload[];
-    comparison: MetricExplorerComparison;
-}> = ({ uniqueEntries, comparison }) => {
-    if (comparison !== MetricExplorerComparison.PREVIOUS_PERIOD) return null;
-
-    if (
-        !uniqueEntries[0]?.payload.metric?.value ||
-        !uniqueEntries[0]?.payload.compareMetric?.value
-    ) {
-        return null;
-    }
-
-    const percentChange =
-        calculateComparisonValue(
-            uniqueEntries[0].payload.metric.value,
-            uniqueEntries[0].payload.compareMetric.value,
-            ComparisonFormatTypes.PERCENTAGE,
-        ) * 100;
-
-    const changeColor =
-        percentChange > 0 ? 'green.7' : percentChange < 0 ? 'red.7' : 'dark.4';
-
-    return (
-        <Group position="right" spacing={4}>
-            <Text c="gray.7" fz={11} fw={500}>
-                Change:
-            </Text>
-            <Text
-                c={changeColor}
-                fz={11}
-                fw={500}
-                ta="right"
-                sx={{ fontFeatureSettings: '"tnum"' }}
-            >
-                {percentChange > 0 ? '+' : ''}
-                {percentChange.toFixed(1)}%
-            </Text>
-        </Group>
-    );
-};
-
-const CustomTooltip = ({
-    active,
-    payload,
-    label,
-    comparison,
-    granularity,
-    is5YearDateRangePreset,
-    dateRange,
-}: CustomTooltipProps & { dateRange?: MetricExplorerDateRange }) => {
-    const hasNoComparison =
-        comparison.comparison === MetricExplorerComparison.NONE;
-    const isSegmented = hasNoComparison && comparison.segmentDimension !== null;
-    const showFullDate =
-        hasNoComparison ||
-        comparison.comparison === MetricExplorerComparison.DIFFERENT_METRIC ||
-        is5YearDateRangePreset;
-
-    const uniqueEntries = useMemo(() => {
-        return uniqBy(payload, getUniqueEntryKey).filter((entry) =>
-            isSegmented && entry.payload.segment
-                ? entry.name === entry.payload.segment
-                : true,
-        );
-    }, [payload, isSegmented]);
-
-    if (!active || !uniqueEntries || !uniqueEntries.length) {
-        return null;
-    }
-
-    const dateLabel = getGranularityLabel(label, granularity, showFullDate);
-    let showDateLabel = false;
-
-    switch (comparison.comparison) {
-        case MetricExplorerComparison.NONE:
-            showDateLabel = true;
-            break;
-        case MetricExplorerComparison.PREVIOUS_PERIOD:
-            showDateLabel =
-                granularity !== TimeFrames.YEAR || is5YearDateRangePreset;
-            break;
-        case MetricExplorerComparison.DIFFERENT_METRIC:
-            showDateLabel = true;
-            break;
-    }
-
-    return (
-        <Stack
-            miw={200}
-            fz={13}
-            fw={500}
-            p="sm"
-            spacing="xs"
-            sx={(theme) => ({
-                backgroundColor: 'white',
-                borderRadius: theme.radius.md,
-                border: `1px solid ${theme.colors.gray[2]}`,
-                boxShadow:
-                    '0px 8px 8px 0px rgba(0, 0, 0, 0.08), 0px 0px 1px 0px rgba(0, 0, 0, 0.25)',
-                flexDirection:
-                    comparison.comparison === MetricExplorerComparison.NONE &&
-                    comparison.segmentDimension === null
-                        ? 'row'
-                        : 'column',
-                justifyContent:
-                    comparison.comparison === MetricExplorerComparison.NONE &&
-                    comparison.segmentDimension === null
-                        ? 'space-between'
-                        : 'flex-start',
-                alignItems:
-                    comparison.comparison === MetricExplorerComparison.NONE &&
-                    comparison.segmentDimension === null
-                        ? 'center'
-                        : 'initial',
-            })}
-        >
-            {showDateLabel && (
-                <>
-                    <Text c="gray.7" fz={13} fw={500}>
-                        {dateLabel}
-                    </Text>
-                    <Divider color="gray.2" />
-                </>
-            )}
-            {uniqueEntries.map((entry) => (
-                <CustomTooltipPayloadEntry
-                    key={getUniqueEntryKey(entry)}
-                    date={label}
-                    is5YearDateRangePreset={is5YearDateRangePreset}
-                    entry={entry}
-                    color={entry.stroke}
-                    comparison={comparison}
-                    dateRange={dateRange}
-                />
-            ))}
-            <PercentageChangeFooter
-                uniqueEntries={uniqueEntries}
-                comparison={comparison.comparison}
-            />
-        </Stack>
-    );
 };
 
 type Props = {
@@ -943,7 +599,7 @@ const MetricsVisualization: FC<Props> = ({
 
                             <RechartsTooltip
                                 content={
-                                    <CustomTooltip
+                                    <MetricExploreTooltip
                                         comparison={query}
                                         granularity={
                                             results.metric.timeDimension
